requirements.txt
# Core dependencies for Chinese Google 5-gram Occupation-Gender-Prestige Project

# Word embeddings
gensim>=4.3.0

# Numerical computing
numpy>=1.24.0

# Data manipulation
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
PyYAML>=6.0

# CLI
fire>=0.5.0

# Web scraping and downloading
requests>=2.31.0
beautifulsoup4>=4.12.0

# Testing
pytest>=7.4.0

# Optional: Jupyter for exploratory analysis
jupyter>=1.0.0
ipykernel>=6.25.0
